tambah folder

+++ backend/controller/kelasController.js
@@ -10,28 +10,23 @@ export const testKelas = asyncHandler(async (req, res) => {
 })
 
 export const CreateKelas = asyncHandler(async (req, res) => {
-    const { nama, keterangan, level, } = req.body;
+    const { nama, keterangan, level, tgllahir, } = req.body;
 
-    const oldNama = await Kelas.findOne({
-        nama: nama
-    });
-    if (!oldNama) {
-        const newKelas = await Kelas.create({
-            nama,
-            keterangan,
-            level,
-            userId: req.user._id
-        })
-        return res.status(200).json({
-            message: "berhasil tambah data kelas",
-        }
-        )
-    } else {
-        return res.status(401).json({
-            message: "Data sudah ada",
-        })
-    }
+    //TIDAK ADA UNIQ COLUMN...............
+    const newKelas = await Kelas.create({
+        nama,
+        keterangan,
+        level,
+        tgllahir,
+        userId: req.user._id
+    })
+    return res.status(200).json({
+        message: "berhasil tambah data kelas",
+        data: newKelas
+    })
 })
+
+
 export const GetAllKelas = asyncHandler(async (req, res) => {
     const KelasData = await Kelas.find()
     return res.status(200).json({
@@ -41,9 +36,9 @@ export const GetAllKelas = asyncHandler(async (req, res) => {
 })
 
 export const GetKelasId = asyncHandler(async (req, res) => {
-    console.log("GetKelasId")
     const idParam = req.params.id
     const detailKelas = await Kelas.findById(idParam)
+    // console.log("disini ..detailKelas.."+detailKelas)
     if (!detailKelas) {
         return res.status(404).json({
             message: "Id Kelas tidak ditemukan"
@@ -56,22 +51,20 @@ export const GetKelasId = asyncHandler(async (req, res) => {
 })
 
 export const GetKelasByName = asyncHandler(async (req, res) => {
-    console.log("GetKelasByName")
-    const kelas = req.params.kelas
-    const detailKelas = await Kelas.find
-    ({ nama: {$regex:kelas} })
-    if (!detailKelas) {
-        return res.status(404).json({
-            message: "Id Kelas tidak ditemukan"
-        })
-    }
-    return res.status(200).json({
-        message: "Document Id kelas berhasil di temukan",
-        data: detailKelas
-    })
+//Untuk disesuaikan pencariannya
+/*const kelas = req.params.kelas
+const detailKelas = await Kelas.find
+({ nama: {$regex:kelas} })
+if (!detailKelas) {
+return res.status(404).json({
+message: "Kelas tidak ditemukan"
 })
-
-
+}
+return res.status(200).json({
+message: "Document Kelas berhasil di temukan", 
+data: detailKelas
+})
+*/})
 export const DeleteKelas = asyncHandler(async (req, res) => {
     //format id harus seuai dengan format ObjectId pad mongoo
     const idParam = req.params.id
@@ -86,7 +79,7 @@ export const DeleteKelas = asyncHandler(async (req, res) => {
             message: "Id pertanyaan tidak ditemukan"
         })
     }
-    //checkPermission(req.user, detailKelas.userId, res)
+    checkPermission(req.user, detailKelas.userId, res)
     const deleteKelas = await Kelas.findByIdAndDelete(idParam)
     return res.status(200).json({
         message: "Delete kelas berhasil"
@@ -94,7 +87,7 @@ export const DeleteKelas = asyncHandler(async (req, res) => {
 })
 
 export const UpdateKelas = asyncHandler(async (req, res) => {
-    const { nama, keterangan, level, } = req.body;
+    const { nama, keterangan, level, tgllahir, } = req.body;
 
     const paramId = req.params.id
     const idKelas = await Kelas.findById(paramId)
@@ -102,10 +95,11 @@ export const UpdateKelas = asyncHandler(async (req, res) => {
         res.status(404)
         throw new Error("Pertanyaan tidak ditemukan")
     }
-    //checkPermission(req.user, idKelas.userId, res)
+    checkPermission(req.user, idKelas.userId, res)
     idKelas.nama = nama
     idKelas.keterangan = keterangan
     idKelas.level = level
+    idKelas.tgllahir = tgllahir
     await idKelas.save()
     return res.status(200).json({
         message: "Berhasil update kelas",
